Organize the structure while maintaining the backward compatibility
