+++ optuna/storages/__init__.py
@@ -5,12 +5,12 @@ from optuna.storages._base import BaseStorage
 from optuna.storages._cached_storage import _CachedStorage
 from optuna.storages._heartbeat import fail_stale_trials
 from optuna.storages._in_memory import InMemoryStorage
+from optuna.storages._journal.backend import JournalFileBackend
+from optuna.storages._journal.backend import JournalFileStorage
 from optuna.storages._journal.base import BaseJournalBackend
 from optuna.storages._journal.base import BaseJournalLogStorage
-from optuna.storages._journal.file import JournalFileBackend
-from optuna.storages._journal.file import JournalFileOpenLock
-from optuna.storages._journal.file import JournalFileStorage
-from optuna.storages._journal.file import JournalFileSymlinkLock
+from optuna.storages._journal.file_lock import JournalFileOpenLock
+from optuna.storages._journal.file_lock import JournalFileSymlinkLock
 from optuna.storages._journal.redis import JournalRedisBackend
 from optuna.storages._journal.redis import JournalRedisStorage
 from optuna.storages._journal.storage import JournalStorage
